Added Octoprime tires
Made some of the cars use Octoprime instead of Carrigan, changed some tests to show that sensor sum works and that's it! It went WAY faster than the Carrigan, because I made Carrigan in such a way that it would be very easy to add another tire. That's the power of TDD. Thank you so much, I had a blast with this, and I hope to get an opportunity to work for Lyft.
Thank you!